Add more detailed error for wrong email
We meant to do this before round 2 but didn’t get to it. We think we
need to help people understand that it’s not their own email address
needed here if they’ve been forwarded the link.

Content is based on questions the first participant in round 2 of
research had with the previous error.